requirements.txt
bottle==0.13.2
Django==5.1.7
docker_py==1.10.6
fastapi==0.115.11
Flask==3.1.0
Jinja2==3.1.3
pandas==2.2.3
plotly==6.0.0
PyYAML==6.0.1
PyYAML==6.0.2
sanic==24.12.0
starlette==0.46.1
streamlit==1.43.1
tornado==6.4
uvicorn==0.34.0
